fix(tests): align fuzz_1h seq counter type with framing.encode signature
`framing.encode(gpa, T, seq_id, msg)` takes a `u32` for `seq_id`
to match `framing.Header.seq_id`'s wire width. The fuzz contexts
declared `sent: u64` and passed it directly to `encode`, which
fails the build on Linux (Zig's compile-error is path-dependent
and macOS's looser implicit integer coercions had been hiding it
locally).

Aligned both fuzz files to the protocol-level type:
- `tests/ipc/fuzz_1h.zig` — `sent: u32`, `recv: u32`.
- `tests/ipc/fuzz_short.zig` — `valid_frames_sent: u32`,
  `valid_frames_recv: u32`.

The wraparound `+%` operator is preserved verbatim; behaviour is
identical to the previous wider counters. The 1 h harness tops out
at ~36 M messages (10 000 msg/s × 3 600 s), well under `u32` max
(~4.3 B), so the narrower type is enough for the post-run sanity
check `recv == sent` too — no separate `u64` stats counter
needed for S6.

Audit of every other `framing.encode` call site:
- `src/core/ipc/connection.zig` (×2) — `next_seq: u32`, passes the
  `u32` `real_seq` derived from it. Correct.
- `tests/ipc/framing.zig` — `framing.encode(gpa, …, 123, &echo)`
  with a comptime literal that fits `u32`. Correct.

`zig build` clean (native), `zig build -Dtarget=x86_64-linux`
clean, `zig build test` exit 0, `zig fmt --check` clean.

+++ tests/ipc/fuzz_1h.zig
@@ -35,8 +35,13 @@ const FuzzCtx = struct {
     server_sock: *ipc.transport.IpcSocket,
     client_sock: *ipc.transport.IpcSocket,
     duration_ms: i64,
-    sent: u64 = 0,
-    recv: u64 = 0,
+    /// Outgoing `seq_id`. Matches the protocol-level `framing.Header.seq_id`
+    /// width (cf. `framing.zig`). 1 h × 10 000 msg/s ≈ 36 M, well under
+    /// `u32` max (~4.3 B), so the wraparound `+%` is theoretical here.
+    sent: u32 = 0,
+    /// Reader-side counter. Same width as `sent` for symmetry —
+    /// drives the post-run sanity check that recv == sent.
+    recv: u32 = 0,
     fault: std.atomic.Value(u8) = std.atomic.Value(u8).init(0),
     stop: std.atomic.Value(u8) = std.atomic.Value(u8).init(0),
 };
